Rename tx handler to tx builder and set up some chain client skeleton
