+++ ocular/Cargo.toml
@@ -55,3 +55,6 @@ hex = "0.3.1"
 cosmrs = "0.5.0"
 rand_core = "0.6.3"
 dirs = "1.0.4"
+cosmos-sdk-proto = { version = "0.10", features = ["grpc"] }
+tendermint-proto = "0.23.5"
+tonic = { version = "0.6", features = ["transport"] }
