+++ ocular/src/chain/client.rs
@@ -9,6 +9,7 @@ use futures::executor;
 use tendermint_rpc;
 
 pub mod query;
+pub mod transactions;
 
 type RpcClient = tendermint_rpc::HttpClient;
