+++ ocular/src/lib.rs
@@ -17,4 +17,4 @@ pub mod cosmos_modules {
 pub mod error;
 pub mod keyring;
 pub mod utils;
-pub mod tx_handler;
+pub mod tx_builder;
